requirements.txt
discord.py==1.7.3
PyNaCl==1.4.0
youtube-dl==2021.6.6
